% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_from_ggplot2.R
\name{autoplot}
\alias{autoplot}
\title{Functions exported from ggplot2}
\description{
autoplot() is exported so that it can be used, even if ggplot2
is not loaded. Even so, we recommend loading ggplot2 using
 \code{library(ggplot2)}.
}
