% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args_score_lurn_si_10.R
\name{check_args_score_lurn_si_10}
\alias{check_args_score_lurn_si_10}
\title{Check the arguments of the score_lurn_si_10}
\usage{
check_args_score_lurn_si_10(input, transfer_vars)
}
\arguments{
\item{input}{is an file supplied by the user containing LURN SI-10 data
(and perhaps other data)}

\item{transfer_vars}{are variables to be included in tehe output}
}
\description{
Check the arguments of the score_lurn_si_10
}
