% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lurn-package.R
\docType{package}
\name{lurn-package}
\alias{lurn}
\alias{lurn-package}
\title{lurn: A package for scoring the LURN SI-10 and LURN SI-29 questionnaires.}
\description{
This package is primarily use for scoring of the LURN SI-10 and LURN SI-29. Additional tools related to LURN research may be added in the future.
}
\section{For more information}{

\url{https://nih-lurn.org/Resources/Questionnaires}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://nih-lurn.org/}
  \item Report bugs at \url{https://github.com/jameswgriffith/lurn/issues}
}

}
\author{
\strong{Maintainer}: James W Griffith \email{j-griffith@northwestern.edu} (\href{https://orcid.org/0000-0002-4840-8692}{ORCID})

}
\keyword{internal}
