% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lurn_si_10_names.R
\name{lurn_si_10_names}
\alias{lurn_si_10_names}
\title{A helper function that returns the recommended variable names of the
LURN SI-10}
\usage{
lurn_si_10_names(include_bother_item = TRUE)
}
\arguments{
\item{include_bother_item}{This is TRUE by default. For the SI-10,
only the 10 symptom questions are scored, so setting this parameter to
FALSE will return only the first 10 item names for symptoms,
and not the item about bother.}
}
\value{
A character vector of recommended variable names for the LURN SI-10.
}
\description{
This function returns a character vector of recommended
variable names for the LURN SI-10. Case matters.
}
\section{For more information}{
 You can check the LURN SI-10 questionnaire
by downloading the pdf at
\url{https://nih-lurn.org/Resources/Questionnaires}.
}

\examples{
\dontrun{
lurn_si_10_names()
lurn_si_10_names(include_bother_item = FALSE)
}
}
