% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lurn_si_29_names.R
\name{lurn_si_29_names}
\alias{lurn_si_29_names}
\title{A helper function that returns the recommended variable names of the
LURN SI-29}
\usage{
lurn_si_29_names(include = c("all", "female", "male"))
}
\arguments{
\item{include}{By default, `all` variable names of the LURN SI-29 are
returned. If desired, the user can request only variable names specific
to `female` respondents or `male` respondents. If needed, the argument
names will be converted to lowercase (i.e., `ALL`, `All`, etc.,
will be silently converted to `all`).}
}
\value{
A character vector of recommended variable names for the LURN SI-29.
}
\description{
This function returns a character vector of recommended
variable names for the LURN SI-29. The user can `include` the complete
list of `all` variable names, or only those variable names specific to
`male` or `female` participants.
}
\examples{
\dontrun{
lurn_si_29_names()

# Only female variable names
lurn_si_29_names(include = "female")

#' Only male variable names
lurn_si_29_names(include = "male")
}
}
