% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_LURN_class.R
\name{remove_LURN_class}
\alias{remove_LURN_class}
\title{Removes the LURN_SI_10 class from an object}
\usage{
remove_LURN_class(x)
}
\arguments{
\item{x}{An object - usually a dataframe that has
been created by score_lurn_si_10.}
}
\value{
x with the LURN_SI_10 class removed.
}
\description{
This function removes the "LURN_SI_10" class
from your object if you do not want to use LURN-specific
plot(), summary(), autoplot() and other methods.
}
\examples{
\dontrun{
some_results <- score_lurn_si_10(input = lurn_si_10_test_data,
output_LURN_class = TRUE)

class(some_results)

some_results <- remove_LURN_class(some_results)

class(some_results)

}
}
\seealso{
score_lurn_si_10()
}
