% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_lurn_si_10.R
\name{score_lurn_si_10}
\alias{score_lurn_si_10}
\title{Returns a score for the SI-10 based on a dataframe of input}
\usage{
score_lurn_si_10(
  input,
  transfer_vars = names(input),
  warn_or_stop = c("warn", "stop")
)
}
\arguments{
\item{input}{A dataframe containing LURN SI-10 items. Other columns may also
be present and will be returned by the function (if desired). The items of
the SI-10 must use the recommended names: SI10_Q1 - SI10_Q10. The input
cannot contain variable names to be returned: lurn_si_10_score,
lurn_si_10_count_valid, or lurn_si_10_note.}

\item{transfer_vars}{A vector of variable names to be found in input.
These variables will be returned in the output along with SI-10 scores}

\item{warn_or_stop}{If set to "warn", warnings will notify the user that
non-numeric or out-of-range data are present. If set to "stop", the
warnings will be printed, but execution will stop. In this case, the user
will need to fix the input data in order to proceed. The default is "warn",
but set to "stop" if you wish to prevent the function from executing when
the input contains any out-of-range or non-numeric data.}
}
\value{
A dataframe of output containing SI-10 scores,
a count of valid items, and a scoring note. Any variables
requested in transfer_vars will also be returned.
}
\description{
This function returns takes a dataframe, extracts the SI-10
items, calculates the SI-10 score, and returns the score along with any
other variables from input requested by the user
}
\details{
If only a subset of variables are desired to be returned,
the column names can be specified in transfer_vars.
It should be noted that out-of-range
values in the items are recoded to NA. Moreover, any character data
will be coerced to numeric or NA as part of the scoring.
}
\examples{
\dontrun{
score_lurn_si_10(input = lurn_data)
}
}
\seealso{
lurn_si_10_names()
}
