% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_lurn_si_29.R
\name{score_lurn_si_29}
\alias{score_lurn_si_29}
\title{Returns scores for the SI-29 based on a dataframe of input}
\usage{
score_lurn_si_29(
  input,
  si_29_names = lurn_si_29_names(),
  gender = "gender",
  gender_levels = c(female = 1, male = 2, NA),
  returned_vars = c("lurn_si_29_total_score", "lurn_si_29_incontinence_score",
    "lurn_si_29_pain_score", "lurn_si_29_voiding_score", "lurn_si_29_urgency_score",
    "lurn_si_29_nocturia_score", "lurn_si_29_bother", "lurn_si_29_note",
    "lurn_si_29_total_count_valid", "lurn_si_29_incontinence_count_valid",
    "lurn_si_29_pain_count_valid", "lurn_si_29_voiding_count_valid",
    "lurn_si_29_urgency_count_valid", "lurn_si_29_nocturia_count_valid"),
  rename_returned_vars_to = NULL,
  transfer_vars = names(input),
  warn_or_stop = c("warn", "stop")
)
}
\arguments{
\item{input}{A dataframe containing LURN SI-29 items. Other columns may also
be present and will be returned by the function (if desired).}

\item{si_29_names}{A vector that identifies the names of the LURN SI-29
items in the "input" data frame. We strongly recommend that you use the
recommended names.}

\item{gender}{A variable that indicates the gender of the participant.
By default this variable name is assumed to be "gender", with numeric
values of "1" for female and "2" for male.}

\item{gender_levels}{By default, gender is assumed to be numerically
coded as  1 = female, 2 = male. We recommend using this coding scheme
in conjunction with this scoring algorithm. If you wish a different
specification, please define a column vector with the female value
first and the male value second, as in:
gender = c(female = 0, male = 1) or gender = c(female = "F", male = "M").
Always specify female first and male second.}

\item{returned_vars}{A vector containing a list of additional scoring
variables that may be desired by the user. The user can request all
possible variables by specifying "all". Individual options are:
\itemize{
 \item{lurn_si_29_total - }{Overall SI-29 score (range 0 to 100)}
 \item{lurn_si_29_incontinence_score - }{Incontinence score (range 0 to 100)}
 \item{lurn_si_29_pain_score - }{Pain score (range 0 to 100)}
 \item{lurn_si_29_voiding_score - }{Voiding difficulty score (range 0 to 100)}
 \item{lurn_si_29_urgency_score - }{Urgency score (range 0 to 100)}
 \item{lurn_si_29_nocturia_score - }{Nocturia score (range 0 to 100)}
 \item{lurn_si_29_bother - }{Overall level of bother by urinary symptoms (same as SI29_Q28}
 \item{lurn_si_29_note - }{Identifies observations with half or more missing items from the total questionnaire or individual subscale}
}}

\item{rename_returned_vars_to}{If desired, the user can specify
the variable names for `returned_vars`.}

\item{transfer_vars}{A vector of variable names to be found in input.
These variables will be returned in the output along with SI-29 scores}

\item{warn_or_stop}{If set to "warn", warnings will notify the user that
non-numeric or out-of-range data are present. If set to "stop", the
warnings will be printed, but execution will stop. In this case, the user
will need to fix the input data in order to proceed.}
}
\value{
A dataframe of output containing SI-29 scores.
}
\description{
This function returns takes a dataframe, extracts the
LURN SI-29 items, calculates the subscale scores, and returns them
along with any other variables from "input" requested by the user
}
\details{
If only a subset of variables are desired, the column names can be
specified in transfer_vars. It should be noted that out-of-range
values in the items are recoded to NA. Moreover, any character data
will be coerced to numeric as part of the scoring.
}
\examples{
\dontrun{
score_lurn_si_29(some_si_29_data)

score_lurn_si_29(input = lurn_si_29_data)
}
}
