% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_lurn_si_10.r
\name{summarize_lurn_si_10}
\alias{summarize_lurn_si_10}
\title{Summarizes the LURN SI-10, using data supplied in a dataframe}
\usage{
summarize_lurn_si_10(input, digits = 1)
}
\arguments{
\item{input}{A dataframe, which
contains LURN SI-10 items. Other columns may also be present.
The items of the SI-10 must use the recommended names:
SI10_Q1-SI10_Q10, and SI10_BOTHER. Case matters for the variable names.}

\item{digits}{How may digits to include in the summary table?
For the proportion of missing data, as least two decimals will be
shown.}
}
\value{
NULL. This function only prints the results.
}
\description{
Summarizes the LURN SI-10, using data supplied in a dataframe
}
\details{
Any dataframe containing the items of the LURN SI-10.
}
\examples{
\dontrun{
summary(lurn_si_10_test_data)
}
}
\seealso{
You can use \code{score_lurn_si_10()} to score the LURN SI-10.
}
