% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.LURN_SI_10.r
\name{summary.LURN_SI_10}
\alias{summary.LURN_SI_10}
\title{Summarizes the LURN SI-10, after it has been scored.}
\usage{
\method{summary}{LURN_SI_10}(object, n_digits = 1, ...)
}
\arguments{
\item{object}{A dataframe, with the additional class of "LURN_SI_10", which
contains LURN SI-10 items and scores. Other columns may also be present.
The items of the SI-10 must use the recommended names:
SI10_Q1-SI10_Q10, and SI10_BOTHER. Case matters for the variable names.}

\item{n_digits}{How may digits to include in the summary table?
For the proportion of missing data, as least two decimals will be
shown.}

\item{...}{Other arguments.}
}
\value{
NULL. This function only prints the results.
}
\description{
Summarizes the LURN SI-10, after it has been scored.
}
\details{
Any dataframe scored using score_lurn_si_10() will be given
a class of "LURN_SI_10", which in turn enables summarization using
this function.
}
\examples{
\dontrun{
summary(lurn_si_10_test_data)
}
}
\seealso{
You can use \code{score_lurn_si_10()} to score the LURN SI-10,
which will return a dataframe with an additional class of "LURN_SI_10".
}
