% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.LURN_SI_10.r
\name{autoplot.LURN_SI_10}
\alias{autoplot.LURN_SI_10}
\title{Uses ggplot2 to create helpful plots for the LURN SI-10.
Plots include a stacked bar chart ("item plot")
for each item of the LURN SI-10
(All 10 symptom items, plus the bother item), as well as a histogram.}
\usage{
\method{autoplot}{LURN_SI_10}(
  object,
  plot_type = c("item plot", "histogram"),
  title = NULL,
  hist_caption_stats = TRUE,
  hist_color = "#7bccc4",
  digits = 1,
  ...
)
}
\arguments{
\item{object}{A dataframe containing LURN SI-10 items. Other columns may
also be present. The items of the SI-10 must use the recommended names:
SI10_Q1-SI10_Q10, and SI10_BOTHER. Case matters for the variable names.}

\item{plot_type}{A parameter used to choose from among built-in
plot types. Choices are "item plot" or "histogram:.}

\item{title}{We encourage you to use a descriptive title for your plot.
This is NULL by default, which will not include a title}

\item{hist_caption_stats}{Logical parameter allowing you to choose
whether you would like LURN SI-10 statistics to be printed in the
caption of your histogram.}

\item{hist_color}{This is set to a pale blue-green by default ("#7bccc4").
If you prefer a black-and-white graph, we recommend "darkgray"}

\item{digits}{The number of decimals to display in the histogram caption.
(set to 1 by default).}

\item{...}{Other arguments passed to other functions.}
}
\value{
A ggplot2 object. Use \code{print()} to display the plot.
}
\description{
Selecting "item plot" (the default)
will create a stacked bar chart
using ggplot2 for each item of
the LURN SI-10 (All 10 symptom items, plus the bother item).
Any responses that are out-of-range or character data will show up in red,
signalling a potential problem in your data. Selecting "histogram"
will create a histogram with intervals of [0, 2), [2, 4) and so on.
}
\details{
Any dataframe scored using score_lurn_si_10() will be given
a class of "LURN_SI_10", allowing autoplot() to be used.
}
\section{Item response coding}{
 Items 1-8 are coded with 0-4;
Items 9, 10 are coded with 0-3;
the bother question is coded with 0-3.
This coding must be respected in order for the plot to be properly produced.
}

\examples{
\dontrun{
autoplot(lurn_si_10_test_data)

# Save the plot
si10_item_plot <- autoplot(lurn_si_10_test_data)
print(si10_item_plot)
plot(si10_item_plot)

si10_item_plot +
ggtitle("LURN data")

}
}
\seealso{
You can use \code{score_lurn_si_10()} to score the LURN SI-10,
which will return a dataframe with an additional class of "LURN_SI_10".
You can use \code{plot()} to automatically print your plot.
}
