% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lurn-package.R
\docType{package}
\name{lurn-package}
\alias{lurn}
\alias{lurn-package}
\title{lurn: Tools for LURN/LURN-II Research: Scoring the LURN SI-10 and LURN SI-29}
\description{
This package is primarily use for scoring of the LURN SI-10 and LURN SI-29. Additional tools related to LURN research may be added in the future.
}
\section{For more information}{

\url{https://nih-lurn.org/Resources/Questionnaires}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://nih-lurn.org/}
  \item Report bugs at \url{https://github.com/jameswgriffith/lurn/issues}
}

}
\keyword{internal}
