% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lurn_internals.R
\docType{data}
\name{lurn_si_10_test_data}
\alias{lurn_si_10_test_data}
\title{Synthetic LURN SI-29 data for testing, based on format from LURN I}
\format{
A dataframe with 13 variables and 7 observations
}
\usage{
lurn_si_10_test_data
}
\description{
Synthetic LURN SI-29 data for testing, based on format from LURN I
}
\keyword{datasets}
