% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_lurn_si_29.R
\name{score_lurn_si_29}
\alias{score_lurn_si_29}
\title{Returns scores for the LURN SI-29 based on a dataframe of input}
\usage{
score_lurn_si_29(
  input,
  transfer_vars = names(input),
  warn_or_stop = c("warn", "stop")
)
}
\arguments{
\item{input}{A dataframe containing LURN SI-29 items. Other columns may also
be present and will be returned by the function (if desired).}

\item{transfer_vars}{A vector of variable names to be found in input.
These variables will be returned in the output along with
the LURN SI-29 scores.}

\item{warn_or_stop}{If set to "warn" (the default),
warnings will notify the user that
non-numeric or out-of-range data are present. If set to "stop", the
warnings will be printed, but execution will stop. In this case, the user
will need to fix the input data in order to proceed.}
}
\value{
A dataframe of output containing SI-29 scores, and any variables
requested in transfer_vars.
}
\description{
This function returns takes a dataframe, extracts the
LURN SI-29 items, calculates the subscale scores, and returns them
along with any other variables from "input" requested by the user.
}
\details{
If only a subset of variables are desired, the column names can be
specified in transfer_vars. It should be noted that out-of-range
values in the items are recoded to NA. Moreover, any character data
will be coerced to numeric as part of the scoring.
}
\section{Setting up your input}{
 Gender
Your input needs to contain a variable \code{Gender} with numeric
values of "1" for female and "2" for male.

 Variable names for the LURN SI-29
You must use the recommended variables names for the LURN SI-29, which
have the form SI29_Q1, SI29_Q2, etc. For item #27, there is SI29_Q27a
and SI29_Q27b. You can use \code{lurn_si_29_names()} to obtain a list of the
recommended variable names.
}

\examples{
\dontrun{
score_lurn_si_29(some_si_29_data)

score_lurn_si_29(input = lurn_si_29_data)
}
}
